refactor(logger): make sanitizeError internal to the logger package
Move sanitizeError (and its GraphQL-specific helpers and types) from
@unique-ag/utils into packages/logger/src/sanitize-error.ts so it is
not part of the public utils API. The function is exclusively used by
the pino err serializer in the logger; exposing it from utils would
invite callers to double-sanitize errors before logging them.

- normalizeError remains in @unique-ag/utils (has external consumers)
- serialize-error-cjs dependency moved from utils to logger
- sanitizeError tests moved to packages/logger/src/sanitize-error.spec.ts
