+++ packages/logger/package.json
@@ -15,7 +15,8 @@
   },
   "dependencies": {
     "@unique-ag/utils": "workspace:*",
-    "remeda": "2.32.0"
+    "remeda": "2.32.0",
+    "serialize-error-cjs": "0.2.0"
   },
   "devDependencies": {
     "@types/node": "22.16.5",
