+++ packages/logger/src/options.ts
@@ -1,6 +1,6 @@
 import { IncomingMessage } from 'node:http';
 import path from 'node:path';
-import { sanitizeError } from '@unique-ag/utils';
+import { sanitizeError } from './sanitize-error';
 import { RequestMethod } from '@nestjs/common';
 import type { Params } from 'nestjs-pino';
 import { isPlainObject } from 'remeda';
