+++ packages/utils/src/index.ts
@@ -5,7 +5,7 @@ export {
   getHttpStatusCodeClass,
   getSlowRequestDurationBucket,
 } from './metrics';
-export { normalizeError, sanitizeError } from './normalize-error';
+export { normalizeError } from './normalize-error';
 export {
   type BatchProcessorOptions,
   processInBatches,
